+++ neuroimg/pipeline/config/localconfig.yaml
@@ -2,15 +2,15 @@
 # home directory that every dataset sits in -> fileutils sets the BIDS structure
 basedatadir:
 #  "/home/ksr/Desktop/contact_localization/efri/"
-#  "/Users/adam2392/Dropbox/phd_research/data/neuroimaging_results/epilepsy/"
-  "/home/adam2392/Documents/Dropbox/phd_research/data/neuroimaging_results/epilepsy/"
+  "/Users/adam2392/Dropbox/phd_research/data/neuroimaging_results/epilepsy/"
+#  "/home/adam2392/Documents/Dropbox/phd_research/data/neuroimaging_results/epilepsy/"
 
 rawdatadir:
 #    "/home/ksr/Desktop/contact_localization/efri/raw_msb/"
 # "/Users/ChesterHuynh/sarmalab/contact_localization_proj/rawdata"
  # "/home/adam2392/hdd2/data/neuroimaging/raw"
-#  "/Users/adam2392/Dropbox/phd_research/data/neuroimaging_results/epilepsy/raw/"
-  "/home/adam2392/Documents/Dropbox/phd_research/data/neuroimaging_results/epilepsy/raw/"
+  "/Users/adam2392/Dropbox/phd_research/data/neuroimaging_results/epilepsy/raw/"
+#  "/home/adam2392/Documents/Dropbox/phd_research/data/neuroimaging_results/epilepsy/raw/"
 ############# define list of files to run this for #############
 patients:
   [
